# api/src/api.yaml
openapi: "3.0.3"
info:
  version: 0.0.1
  title: Micro
  license:
    name: MIT
servers:
  - url: http://localhost:8000
paths:
  /delegation:
    parameters:
      - name: usrName
        in: header
        required: true
        schema:
          type: string
      - name: usrId
        in: header
        required: true
        schema:
          type: string
      - name: usrRole
        in: header
        required: true
        schema:
          type: string
    post:
      summary: creat delegation
      operationId: creatDelegation
      tags:
        - delegation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/creatDelegationRequest"
        required: true
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                type: string
        '400':
          description: error
          content:
            application/json:
              schema:
                type: string
  /delegation/{id}/files:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: usrName
        in: header
        required: true
        schema:
          type: string
      - name: usrId
        in: header
        required: true
        schema:
          type: string
      - name: usrRole
        in: header
        required: true
        schema:
          type: string
    post: #finished
      operationId: createDelegationFile
      tags:
        - delegation
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                # The property name 'file' will be used for all files.
                usrManual:
                  type: string
                  format: binary
                installationManual:
                  type: string
                  format: binary
                operationManual:
                  type: string
                  format: binary
                maintenanceManual:
                  type: string
                  format: binary
      responses:
        '201':
          description: created
        '404':
          description: not found delegation
    get:
      operationId: listDelegationFile
      tags:
        - delegations
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/allFiles'
        "404":
          description: delegation not found
  /delegation/{id}:
    parameters:
      - name: usrName
        in: header
        required: true
        schema:
          type: string
      - name: usrId
        in: header
        required: true
        schema:
          type: string
      - name: usrRole
        in: header
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        description: The id of delegation
        schema:
          type: string
    get:
      operationId: findDelegation
      tags:
        - delegation
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/delegationItem"
        '404':
          description: delegation not found
    delete:
      operationId: deleteDelegation
      tags:
        - delegation
      responses:
        '200':
          description: ok
        '403':
          description: bad permission
        '404':
          description: delegation not found
  /delegation/{id}/applicationTable:
    parameters:
      - name: id
        in: path
        required: true
        description: The id of delegation
        schema:
          type: string
    put:
      operationId: updateApplicationTable
      tags:
        - delegation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/delegationApplicationTable"
        required: true
      responses:
        '200':
          description: update succeed
        '404':
          description: delegation not found
        '403':
          description: delegation can no longer be modified
  /delegation/{id}/functionTable:
    parameters:
      - name: id
        in: path
        required: true
        description: The id of delegation
        schema:
          type: string
    put:
      operationId: updateFuntionTable
      tags:
        - delegation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/delegationFunctionTable"
        required: true
      responses:
        '200':
          description: update succeed
        '404':
          description: delegation not found
        '403':
          description: delegation can no longer be modified
  /delegation/{id}/fileListTable:
    parameters:
      - name: id
        in: path
        required: true
        description: The id of delegation
        schema:
          type: string
    put:
      operationId: updatefileListTable
      tags:
        - delegation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/delegationFileList"
        required: true
      responses:
        '200':
          description: update succeed
        '404':
          description: delegation not found
        '403':
          description: delegation can no longer be modified

  /delegations:
    parameters:
      - name: usrName
        in: header
        required: true
        schema:
          type: string
      - name: usrId
        in: header
        required: true
        schema:
          type: string
      - name: usrRole
        in: header
        required: true
        schema:
          type: string
    get:
      operationId: listDelegations
      tags:
        - delegation
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/delegationItem"

        '404':
          description: usr not found
  /audit/delegation/{id}:
    parameters:
      - name: usrName
        in: header
        required: true
        schema:
          type: string
      - name: usrId
        in: header
        required: true
        schema:
          type: string
      - name: usrRole
        in: header
        required: true
        schema:
          type: string
      - name: id
        in: path
        required: true
        description: The id of delegation
        schema:
          type: string
    post:
      operationId: auditDelegation
      tags:
        - delegation
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/delegationAuditResult"
        required: true
      responses:
        '200':
          description: ok
        '404':
          description: delegation not found
        '403':
          description: bad permission
        '423':
          description: has been audited
components:
  schemas:
    delegationApplicationTable:
      type: object
      properties:
        name:
          type: string
        测试类型(其他):
          type: string
        测试类型:
          type: array
          items:
            type: string
        软件名称:
          type: string
        版本号:
          type: string
        委托单位(中文):
          type: string
        委托单位(英文):
          type: string
        开发单位:
          type: string
        单位性质:
          type: string
        软件用户对象描述:
          type: string
        主要功能及用途简介(限300字):
          type: string
        测试依据:
          type: array
          items:
            type: string
        需要测试的技术指标:
          type: array
          items:
            type: string
        软件规模:
          $ref: "#/components/schemas/softwareScale"
        软件类型:
          type: string
        运行环境:
          $ref: "#/components/schemas/runtimeEnvironment"
        样品和数量:
          $ref: "#/components/schemas/sampleAndQuantity"
        希望测试完成时间:
          type: string
        委托单位信息:
          $ref: "#/components/schemas/clientEnterpriseInfo"

    delegationFunctionTable:
      type: object
      properties:
        name:
          type: string
    delegationFileList:
      type: object
      properties:
        fileName:
          type: array
          items:
            type: string
    delegationAuditResult:
      type: object
      properties:
        suggestion:
          type: string
        state:
          type: string
    delegationItem:
      type: object
      properties:
        usrBelonged:
          type: string
        delegationId:
          type: string
        state:
          type: string
        suggestion:
          type: string
        applicationTable:
          $ref: "#/components/schemas/delegationApplicationTable"
        functionTable:
          $ref: "#/components/schemas/delegationFunctionTable"
        fileList:
          $ref: "#/components/schemas/delegationFileList"
    creatDelegationRequest:
      type: object
      properties:
        applicationTable:
          $ref: "#/components/schemas/delegationApplicationTable"
        functionTable:
          $ref: "#/components/schemas/delegationFunctionTable"
        fileList:
          $ref: "#/components/schemas/delegationFileList"


    softwareScale:
      type: object
      properties:
        功能数(到最后一级菜单):
          type: string

    runtimeEnvironment:
      type: object
      properties:
        客户端:
          $ref: "#/components/schemas/clientRuntimeEnvironment"
        服务器端:
          $ref: "#/components/schemas/serverRuntimeEnvironment"
        网络环境:
          type: string

    clientRuntimeEnvironment:
      type: object
      properties:
        操作系统-Windows版本:
          type: string
        操作系统:
          type: array
          items:
            type: string
        内存要求:
          type: string
        其他要求:
          type: string

    serverRuntimeEnvironment:
      type: object
      properties:
        硬件:
          $ref: "#/components/schemas/hardwareEnvironment"
        软件:
          $ref: "#/components/schemas/softwareEnvironment"


    hardwareEnvironment:
      type: object
      properties:
        架构:
          type: array
          items:
            type: string
        内存要求:
          type: string
        硬盘要求:
          type: string
        其他要求:
          type: string

    softwareEnvironment:
      type: object
      properties:
        操作系统:
          type: string
        版本:
          type: string
        编程语言:
          type: string
        架构:
          type: array
          items:
            type: string
        数据库:
          type: string
        中间件:
          type: string
        其他支撑软件:
          type: string

    sampleAndQuantity:
      type: object
      properties:
        软件介质:
          type: string
        提交的样品(硬拷贝资料及硬件)五年保存期满后:
          type: string

    clientEnterpriseInfo:
      type: object
      properties:
        电话:
          type: string
        传真:
          type: string
        地址:
          type: string
        邮编:
          type: string
        联系人:
          type: string
        手机:
          type: string
        E-mail:
          type: string
        网址:
          type: string


    singleFile:
      type: object
      properties:
        fileName:
          type: string
        fileType:
          type: string
        fileUri:
          type: string
    allFiles:
      type: object
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/singleFile'